+++ packages/flutter_image/lib/network.dart
@@ -17,6 +17,9 @@ import 'dart:ui' as ui;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
+// Method signature for _loadWithRetry decode callbacks.
+typedef _SimpleDecoderCallback = Future<ui.Codec> Function(ui.ImmutableBuffer buffer);
+
 /// Fetches the image from the given URL, associating it with the given scale.
 ///
 /// If [fetchStrategy] is specified, uses it instead of the
@@ -95,8 +98,11 @@ class NetworkImageWithRetry extends ImageProvider<NetworkImageWithRetry> {
   }
 
   @override
-  ImageStreamCompleter loadImage(
-      NetworkImageWithRetry key, ImageDecoderCallback decode) {
+  ImageStreamCompleter loadBuffer(
+      // TODO(cyanglaz): migrate to use the new APIs
+      // https://github.com/flutter/flutter/issues/105336
+      // ignore: deprecated_member_use
+      NetworkImageWithRetry key, DecoderBufferCallback decode) {
     return OneFrameImageStreamCompleter(_loadWithRetry(key, decode),
         informationCollector: () sync* {
       yield ErrorDescription('Image provider: $this');
@@ -124,7 +130,7 @@ class NetworkImageWithRetry extends ImageProvider<NetworkImageWithRetry> {
   }
 
   Future<ImageInfo> _loadWithRetry(
-      NetworkImageWithRetry key, ImageDecoderCallback decode) async {
+      NetworkImageWithRetry key, _SimpleDecoderCallback decode) async {
     assert(key == this);
 
     final Stopwatch stopwatch = Stopwatch()..start();